Factory Method
In Factory pattern, we create object without exposing the creation
logic. In this pattern, an interface is used for creating an object, but
let subclass decide which class to instantiate. The creation of object
is done when it is required. The Factory method allows a class later
instantiation to subclasses.