Adding Prototype Pattern
Prototype pattern is used to create a duplicate object or clone of the
current object to enhance performance. This pattern is used when
creation of object is costly or complex.
For Example: An object is to be created after a costly database
operation. We can cache the object, returns its clone on next request
and update the database as and when needed thus reducing database calls.