Bridge Pattern
Bridge pattern is used to separate an abstraction from its
implementation so that both can be modified independently.
This pattern involves an interface which acts as a bridge between the
abstraction class and implementer classes and also makes the
functionality of implementer class independent from the abstraction
class. Both types of classes can be modified without affecting to each
other.