Facade Pattern
Facade pattern hides the complexities of the system and provides an
interface to the client using which the client can access the system.
This pattern involves a single wrapper class which contains a set of
members which are required by client. These members access the system on
behalf of the facade client and hide the implementation details